requirements.txt
torch>=2.2.0
torchvision>=0.17.0
tensorboard>=2.15.0
numpy>=1.26.0
Pillow>=10.0.0
tqdm>=4.66.0
